azure-functions/requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b4
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.12.0

requests2>=2.16.0
arcgis>=1.5.3
python-dotenv>=0.5.1
pillow
